refactor: siteData.tsをconstants.tsに統合
- SITE_METADATAとSiteMetadata型をsrc/constants.tsに移動
- 既存のGA_TRACKING_IDと同じファイルで管理
- インポートパスを@/constants/siteDataから@/constantsに更新
- 不要になったsrc/constants/siteData.tsを削除
